+++ src/Console/TubePeekCommand.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Phlib\Beanstalk\Console;
 
 use Phlib\Beanstalk\Exception\InvalidArgumentException;
+use Phlib\Beanstalk\Exception\NotFoundException;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -46,25 +47,24 @@ class TubePeekCommand extends AbstractCommand
         $this->getBeanstalk()
             ->useTube($input->getArgument('tube'));
 
-        // Use switch instead of `->{'peek' . $status}` to allow static analysis
-        switch ($status) {
-            case 'ready':
-                $job = $this->getBeanstalk()->peekReady();
-                break;
-            case 'delayed':
-                $job = $this->getBeanstalk()->peekDelayed();
-                break;
-            case 'buried':
-                $job = $this->getBeanstalk()->peekBuried();
-                break;
-            default:
-                throw new InvalidArgumentException("Specified status '{$status}' is not valid.");
-        }
-
-        if ($job === null) {
-            $output->writeln("No jobs found in '{$status}' status.");
-        } else {
+        try {
+            // Use switch instead of `->{'peek' . $status}` to allow static analysis
+            switch ($status) {
+                case 'ready':
+                    $job = $this->getBeanstalk()->peekReady();
+                    break;
+                case 'delayed':
+                    $job = $this->getBeanstalk()->peekDelayed();
+                    break;
+                case 'buried':
+                    $job = $this->getBeanstalk()->peekBuried();
+                    break;
+                default:
+                    throw new InvalidArgumentException("Specified status '{$status}' is not valid");
+            }
             $this->displayJob($job, $output);
+        } catch (NotFoundException $e) {
+            $output->writeln("No jobs found in '{$status}' status.");
         }
 
         return 0;